internship_bot/requirements.txt
requests>=2.32.0
PyYAML>=6.0.1
gspread>=6.1.2
google-auth>=2.35.0
python-dateutil>=2.9.0.post0
